- Add Download chip to artists and albums - Break up DatabaseDao into smaller interfaces
