+++ app/src/main/java/com/dd3boh/outertune/db/DatabaseDao.kt
@@ -6,21 +6,18 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.RawQuery
-import androidx.room.RewriteQueriesToDropUnusedColumns
 import androidx.room.Transaction
 import androidx.room.Update
 import androidx.room.Upsert
-import androidx.sqlite.db.SimpleSQLiteQuery
 import androidx.sqlite.db.SupportSQLiteQuery
-import com.dd3boh.outertune.constants.AlbumSortType
-import com.dd3boh.outertune.constants.ArtistSongSortType
-import com.dd3boh.outertune.constants.ArtistSortType
-import com.dd3boh.outertune.constants.PlaylistSortType
-import com.dd3boh.outertune.constants.SongSortType
+import com.dd3boh.outertune.db.daos.AlbumsDao
+import com.dd3boh.outertune.db.daos.ArtistsDao
+import com.dd3boh.outertune.db.daos.PlaylistsDao
+import com.dd3boh.outertune.db.daos.QueueDao
+import com.dd3boh.outertune.db.daos.SongsDao
 import com.dd3boh.outertune.db.entities.Album
 import com.dd3boh.outertune.db.entities.AlbumArtistMap
 import com.dd3boh.outertune.db.entities.AlbumEntity
-import com.dd3boh.outertune.db.entities.AlbumWithSongs
 import com.dd3boh.outertune.db.entities.Artist
 import com.dd3boh.outertune.db.entities.ArtistEntity
 import com.dd3boh.outertune.db.entities.Event
@@ -28,10 +25,7 @@ import com.dd3boh.outertune.db.entities.EventWithSong
 import com.dd3boh.outertune.db.entities.FormatEntity
 import com.dd3boh.outertune.db.entities.GenreEntity
 import com.dd3boh.outertune.db.entities.LyricsEntity
-import com.dd3boh.outertune.db.entities.PlayCountEntity
 import com.dd3boh.outertune.db.entities.Playlist
-import com.dd3boh.outertune.db.entities.PlaylistEntity
-import com.dd3boh.outertune.db.entities.PlaylistSong
 import com.dd3boh.outertune.db.entities.PlaylistSongMap
 import com.dd3boh.outertune.db.entities.QueueEntity
 import com.dd3boh.outertune.db.entities.QueueSongMap
@@ -42,186 +36,20 @@ import com.dd3boh.outertune.db.entities.SongAlbumMap
 import com.dd3boh.outertune.db.entities.SongArtistMap
 import com.dd3boh.outertune.db.entities.SongEntity
 import com.dd3boh.outertune.db.entities.SongGenreMap
-import com.dd3boh.outertune.extensions.reversed
 import com.dd3boh.outertune.extensions.toSQLiteQuery
 import com.dd3boh.outertune.models.MediaMetadata
 import com.dd3boh.outertune.models.MultiQueueObject
-import com.dd3boh.outertune.models.QueueBoard
 import com.dd3boh.outertune.models.toMediaMetadata
-import com.dd3boh.outertune.ui.utils.resize
 import com.zionhuang.innertube.models.AlbumItem
-import com.zionhuang.innertube.models.PlaylistItem
 import com.zionhuang.innertube.models.SongItem
 import com.zionhuang.innertube.pages.AlbumPage
-import com.zionhuang.innertube.pages.ArtistPage
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.sync.withLock
-import java.time.LocalDateTime
-import java.time.ZoneOffset
 
 @Dao
-interface DatabaseDao {
-    @Transaction
-    @Query("SELECT * FROM song WHERE inLibrary IS NOT NULL ORDER BY rowId")
-    fun songsByRowIdAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE inLibrary IS NOT NULL ORDER BY inLibrary")
-    fun songsByCreateDateAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE inLibrary IS NOT NULL ORDER BY date")
-    fun songsByReleaseDateAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE inLibrary IS NOT NULL ORDER BY dateModified")
-    fun songsByDateModifiedAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE inLibrary IS NOT NULL ORDER BY title COLLATE NOCASE ASC")
-    fun songsByNameAsc(): Flow<List<Song>>
+interface DatabaseDao : SongsDao, AlbumsDao, ArtistsDao, PlaylistsDao, QueueDao {
 
     @Transaction
     @Query("""
-        SELECT * FROM song 
-        WHERE inLibrary IS NOT NULL 
-        ORDER BY (
-            SELECT LOWER(GROUP_CONCAT(name, ''))
-            FROM artist
-            WHERE id IN (SELECT artistId FROM song_artist_map WHERE songId = song.id)
-            ORDER BY name
-        ) COLLATE NOCASE
-    """)
-    fun songsByArtistAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE inLibrary IS NOT NULL ORDER BY totalPlayTime")
-    fun songsByPlayTimeAsc(): Flow<List<Song>>
-
-    @Transaction
-    @RewriteQueriesToDropUnusedColumns
-    @Query("""
-        SELECT song.*, (SELECT SUM(playCount.count) 
-            FROM playCount 
-            WHERE playCount.song = song.id) AS pc 
-        FROM song 
-        WHERE inLibrary IS NOT NULL 
-        ORDER BY pc ASC
-    """)
-    fun songsByPlayCountAsc(): Flow<List<Song>>
-
-    fun songs(sortType: SongSortType, descending: Boolean) =
-        when (sortType) {
-            SongSortType.CREATE_DATE -> songsByCreateDateAsc()
-            SongSortType.MODIFIED_DATE -> songsByDateModifiedAsc()
-            SongSortType.RELEASE_DATE -> songsByReleaseDateAsc()
-            SongSortType.NAME -> songsByNameAsc()
-            SongSortType.ARTIST -> songsByArtistAsc()
-            SongSortType.PLAY_TIME -> songsByPlayTimeAsc()
-        }.map { it.reversed(descending) }
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE liked ORDER BY rowId")
-    fun likedSongsByRowIdAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE liked ORDER BY likedDate")
-    fun likedSongsByCreateDateAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE liked ORDER BY date")
-    fun likedSongsByReleaseDateAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE liked ORDER BY dateModified")
-    fun likedSongsByDateModifiedAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE liked ORDER BY title COLLATE NOCASE ASC")
-    fun likedSongsByNameAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("""
-        SELECT * FROM song 
-        WHERE liked 
-        ORDER BY (
-            SELECT LOWER(GROUP_CONCAT(name, ''))
-            FROM artist
-            WHERE id IN (SELECT artistId FROM song_artist_map WHERE songId = song.id)
-            ORDER BY name
-        ) COLLATE NOCASE
-    """)
-    fun likedSongsByArtistAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE liked ORDER BY totalPlayTime")
-    fun likedSongsByPlayTimeAsc(): Flow<List<Song>>
-
-    @Transaction
-    @RewriteQueriesToDropUnusedColumns
-    @Query("""
-        SELECT song.*, (SELECT SUM(playCount.count) 
-            FROM playCount 
-            WHERE playCount.song = song.id) AS pc 
-        FROM song 
-        WHERE liked IS NOT NULL 
-        ORDER BY pc ASC
-    """)
-    fun likedSongsByPlayCountAsc(): Flow<List<Song>>
-
-    fun likedSongs(sortType: SongSortType, descending: Boolean) =
-        when (sortType) {
-            SongSortType.CREATE_DATE -> likedSongsByCreateDateAsc()
-            SongSortType.MODIFIED_DATE -> likedSongsByDateModifiedAsc()
-            SongSortType.RELEASE_DATE -> ikedSongsByReleaseDateAsc()
-            SongSortType.NAME -> likedSongsByNameAsc()
-            SongSortType.ARTIST -> likedSongsByArtistAsc()
-            SongSortType.PLAY_TIME -> likedSongsByPlayTimeAsc()
-        }.map { it.reversed(descending) }
-
-    @Query("SELECT COUNT(1) FROM song WHERE liked")
-    fun likedSongsCount(): Flow<Int>
-
-    @Transaction
-    @Query("SELECT song.* FROM song JOIN song_album_map ON song.id = song_album_map.songId WHERE song_album_map.albumId = :albumId")
-    fun albumSongs(albumId: String): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM playlist_song_map WHERE playlistId = :playlistId ORDER BY position")
-    fun playlistSongs(playlistId: String): Flow<List<PlaylistSong>>
-
-    @Transaction
-    @Query("SELECT song.* FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = :artistId AND inLibrary IS NOT NULL ORDER BY inLibrary")
-    fun artistSongsByCreateDateAsc(artistId: String): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT song.* FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = :artistId AND inLibrary IS NOT NULL ORDER BY title COLLATE NOCASE ASC")
-    fun artistSongsByNameAsc(artistId: String): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT song.* FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = :artistId AND inLibrary IS NOT NULL ORDER BY totalPlayTime")
-    fun artistSongsByPlayTimeAsc(artistId: String): Flow<List<Song>>
-
-    fun artistSongs(artistId: String, sortType: ArtistSongSortType, descending: Boolean) =
-        when (sortType) {
-            ArtistSongSortType.CREATE_DATE -> artistSongsByCreateDateAsc(artistId)
-            ArtistSongSortType.NAME -> artistSongsByNameAsc(artistId)
-            ArtistSongSortType.PLAY_TIME -> artistSongsByPlayTimeAsc(artistId)
-        }.map { it.reversed(descending) }
-
-    @Transaction
-    @Query("SELECT song.* FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = :artistId AND inLibrary IS NOT NULL LIMIT :previewSize")
-    fun artistSongsPreview(artistId: String, previewSize: Int = 3): Flow<List<Song>>
-
-    @Transaction
-    @Query(
-        """
         SELECT song.*
         FROM (SELECT *, COUNT(1) AS referredCount
               FROM related_song_map
@@ -248,426 +76,14 @@ interface DatabaseDao {
                                LIMIT 10))
         ORDER BY referredCount DESC
         LIMIT 100
-    """
-    )
-    fun quickPicks(now: Long = System.currentTimeMillis()): Flow<List<Song>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT *
-        FROM song
-        WHERE id IN (SELECT songId
-                     FROM event
-                     WHERE timestamp > :fromTimeStamp
-                     GROUP BY songId
-                     ORDER BY SUM(playTime) DESC
-                     LIMIT :limit)
-    """
-    )
-    fun mostPlayedSongs(fromTimeStamp: Long, limit: Int = 6): Flow<List<Song>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT artist.*,
-               (SELECT COUNT(1)
-                FROM song_artist_map
-                         JOIN song ON song_artist_map.songId = song.id
-                WHERE artistId = artist.id
-                  AND song.inLibrary IS NOT NULL) AS songCount
-        FROM artist
-                 JOIN(SELECT artistId, SUM(songTotalPlayTime) AS totalPlayTime
-                      FROM song_artist_map
-                               JOIN (SELECT songId, SUM(playTime) AS songTotalPlayTime
-                                     FROM event
-                                     WHERE timestamp > :fromTimeStamp
-                                     GROUP BY songId) AS e
-                                    ON song_artist_map.songId = e.songId
-                      GROUP BY artistId
-                      ORDER BY totalPlayTime DESC
-                      LIMIT :limit)
-                     ON artist.id = artistId
-    """
-    )
-    fun mostPlayedArtists(fromTimeStamp: Long, limit: Int = 6): Flow<List<Artist>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT album.*
-        FROM album
-                 JOIN(SELECT albumId
-                      FROM song
-                               JOIN (SELECT songId, SUM(playTime) AS songTotalPlayTime
-                                     FROM event
-                                     WHERE timestamp > :fromTimeStamp
-                                     GROUP BY songId) AS e
-                                    ON song.id = e.songId
-                      WHERE albumId IS NOT NULL
-                      GROUP BY albumId
-                      ORDER BY SUM(songTotalPlayTime) DESC
-                      LIMIT :limit)
-                     ON album.id = albumId
-    """
-    )
-    fun mostPlayedAlbums(fromTimeStamp: Long, limit: Int = 6): Flow<List<Album>>
-
-    @Query("SELECT sum(count) from playCount WHERE song = :songId")
-    fun getLifetimePlayCount(songId: String?): Flow<Int>
-    @Query("SELECT sum(count) from playCount WHERE song = :songId AND year = :year")
-    fun getPlayCountByYear(songId: String?, year: Int): Flow<Int>
-    @Query("SELECT count from playCount WHERE song = :songId AND year = :year AND month = :month")
-    fun getPlayCountByMonth(songId: String?, year: Int, month: Int): Flow<Int>
-
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE id = :songId")
-    fun song(songId: String?): Flow<Song?>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE dateDownload IS NOT NULL ORDER BY dateDownload")
-    fun downloadNoLocalSongs(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE isLocal OR dateDownload IS NOT NULL ORDER BY inLibrary")
-    fun downloadSongsByCreateDateAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE isLocal OR dateDownload IS NOT NULL ORDER BY date")
-    fun downloadSongsByReleaseDateAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE isLocal OR dateDownload IS NOT NULL ORDER BY dateModified")
-    fun downloadSongsByDateModifiedAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE isLocal OR dateDownload IS NOT NULL ORDER BY title COLLATE NOCASE ASC")
-    fun downloadSongsByNameAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("""
-        SELECT * FROM song
-        WHERE isLocal OR dateDownload IS NOT NULL
-        ORDER BY (
-            SELECT LOWER(GROUP_CONCAT(name, ''))
-            FROM artist
-            WHERE id IN (SELECT artistId FROM song_artist_map WHERE songId = song.id)
-            ORDER BY name
-        ) COLLATE NOCASE
     """)
-    fun downloadSongsByArtistAsc(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE isLocal OR dateDownload IS NOT NULL ORDER BY totalPlayTime")
-    fun downloadSongsByPlayTimeAsc(): Flow<List<Song>>
-
-    fun downloadSongs(sortType: SongSortType, descending: Boolean) =
-        when (sortType) {
-            SongSortType.CREATE_DATE -> downloadSongsByCreateDateAsc()
-            SongSortType.MODIFIED_DATE -> downloadSongsByDateModifiedAsc()
-            SongSortType.RELEASE_DATE -> downloadSongsByReleaseDateAsc()
-            SongSortType.NAME -> downloadSongsByNameAsc()
-            SongSortType.ARTIST -> downloadSongsByArtistAsc()
-            SongSortType.PLAY_TIME -> downloadSongsByPlayTimeAsc()
-        }.map { it.reversed(descending) }
-
-    @Transaction
-    @Query("SELECT * FROM song WHERE isLocal = 1 and inLibrary IS NOT NULL")
-    fun allLocalSongs(): Flow<List<Song>>
-
-    @Transaction
-    @Query("SELECT * FROM artist WHERE isLocal != 1")
-    fun allRemoteArtists(): Flow<List<ArtistEntity>>
-
-    @Transaction
-    @Query("SELECT * FROM artist WHERE isLocal = 1")
-    fun allLocalArtists(): Flow<List<ArtistEntity>>
+    fun quickPicks(now: Long = System.currentTimeMillis()): Flow<List<Song>>
 
     @Query("SELECT * FROM format WHERE id = :id")
     fun format(id: String?): Flow<FormatEntity?>
 
     @Query("SELECT * FROM lyrics WHERE id = :id")
     fun lyrics(id: String?): Flow<LyricsEntity?>
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE songCount > 0 ORDER BY rowId")
-    fun artistsByCreateDateAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE songCount > 0 ORDER BY name COLLATE NOCASE ASC")
-    fun artistsByNameAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE songCount > 0 ORDER BY songCount")
-    fun artistsBySongCountAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT artist.*,
-               (SELECT COUNT(1)
-                FROM song_artist_map
-                         JOIN song ON song_artist_map.songId = song.id
-                WHERE artistId = artist.id
-                  AND song.inLibrary IS NOT NULL) AS songCount
-        FROM artist
-                 JOIN(SELECT artistId, SUM(totalPlayTime) AS totalPlayTime
-                      FROM song_artist_map
-                               JOIN song
-                                    ON song_artist_map.songId = song.id
-                      GROUP BY artistId
-                      ORDER BY totalPlayTime)
-                     ON artist.id = artistId
-        WHERE songCount > 0
-    """
-    )
-    fun artistsByPlayTimeAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE bookmarkedAt IS NOT NULL ORDER BY bookmarkedAt")
-    fun artistsBookmarkedByCreateDateAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE bookmarkedAt IS NOT NULL ORDER BY name COLLATE NOCASE ASC")
-    fun artistsBookmarkedByNameAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE bookmarkedAt IS NOT NULL ORDER BY songCount")
-    fun artistsBookmarkedBySongCountAsc(): Flow<List<Artist>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT artist.*,
-               (SELECT COUNT(1)
-                FROM song_artist_map
-                         JOIN song ON song_artist_map.songId = song.id
-                WHERE artistId = artist.id
-                  AND song.inLibrary IS NOT NULL) AS songCount
-        FROM artist
-                 JOIN(SELECT artistId, SUM(totalPlayTime) AS totalPlayTime
-                      FROM song_artist_map
-                               JOIN song
-                                    ON song_artist_map.songId = song.id
-                      GROUP BY artistId
-                      ORDER BY totalPlayTime)
-                     ON artist.id = artistId
-        WHERE bookmarkedAt IS NOT NULL
-    """
-    )
-    fun artistsBookmarkedByPlayTimeAsc(): Flow<List<Artist>>
-
-    fun artists(sortType: ArtistSortType, descending: Boolean) =
-        when (sortType) {
-            ArtistSortType.CREATE_DATE -> artistsByCreateDateAsc()
-            ArtistSortType.NAME -> artistsByNameAsc()
-            ArtistSortType.SONG_COUNT -> artistsBySongCountAsc()
-            ArtistSortType.PLAY_TIME -> artistsByPlayTimeAsc()
-        }.map { artists ->
-            artists
-                .filter { it.artist.isYouTubeArtist || it.artist.isLocalArtist } // temp: add ui to filter by local or remote or something idk
-                .reversed(descending)
-        }
-
-    fun artistsBookmarked(sortType: ArtistSortType, descending: Boolean) =
-        when (sortType) {
-            ArtistSortType.CREATE_DATE -> artistsBookmarkedByCreateDateAsc()
-            ArtistSortType.NAME -> artistsBookmarkedByNameAsc()
-            ArtistSortType.SONG_COUNT -> artistsBookmarkedBySongCountAsc()
-            ArtistSortType.PLAY_TIME -> artistsBookmarkedByPlayTimeAsc()
-        }.map { artists ->
-            artists
-                .filter { it.artist.isYouTubeArtist }
-                .reversed(descending)
-        }
-
-    @Query("SELECT *, (SELECT COUNT(1) FROM song_artist_map JOIN song ON song_artist_map.songId = song.id WHERE artistId = artist.id AND song.inLibrary IS NOT NULL) AS songCount FROM artist WHERE id = :id")
-    fun artist(id: String): Flow<Artist?>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL) ORDER BY rowId")
-    fun albumsByCreateDateAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL) ORDER BY title COLLATE NOCASE ASC")
-    fun albumsByNameAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("""
-        SELECT * FROM album
-        WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL)
-        ORDER BY (
-            SELECT LOWER(GROUP_CONCAT(name, ''))
-            FROM artist
-            WHERE id IN (SELECT artistId FROM album_artist_map WHERE albumId = album.id)
-            ORDER BY name
-        ) COLLATE NOCASE
-    """)
-    fun albumByArtistAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL) ORDER BY year")
-    fun albumsByYearAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL) ORDER BY songCount")
-    fun albumsBySongCountAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL) ORDER BY duration")
-    fun albumsByLengthAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT album.*
-        FROM album
-                 JOIN song
-                      ON song.albumId = album.id
-        WHERE EXISTS(SELECT * FROM song WHERE song.albumId = album.id AND song.inLibrary IS NOT NULL)
-        GROUP BY album.id
-        ORDER BY SUM(song.totalPlayTime)
-    """
-    )
-    fun albumsByPlayTimeAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE bookmarkedAt IS NOT NULL ORDER BY rowId")
-    fun albumsLikedByCreateDateAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE bookmarkedAt IS NOT NULL ORDER BY title COLLATE NOCASE ASC")
-    fun albumsLikedByNameAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("""
-        SELECT * FROM album
-        WHERE bookmarkedAt IS NOT NULL
-        ORDER BY (
-            SELECT LOWER(GROUP_CONCAT(name, ''))
-            FROM artist
-            WHERE id IN (SELECT artistId FROM album_artist_map WHERE albumId = album.id)
-            ORDER BY name
-        ) COLLATE NOCASE
-    """)
-    fun albumLikeByArtistAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE bookmarkedAt IS NOT NULL ORDER BY year")
-    fun albumsLikedByYearAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE bookmarkedAt IS NOT NULL ORDER BY songCount")
-    fun albumsLikedBySongCountAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE bookmarkedAt IS NOT NULL ORDER BY duration")
-    fun albumsLikedByLengthAsc(): Flow<List<Album>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT album.*
-        FROM album
-                 JOIN song
-                      ON song.albumId = album.id
-        WHERE bookmarkedAt IS NOT NULL
-        GROUP BY album.id
-        ORDER BY SUM(song.totalPlayTime)
-    """
-    )
-    fun albumsLikedByPlayTimeAsc(): Flow<List<Album>>
-
-    fun albums(sortType: AlbumSortType, descending: Boolean) =
-        when (sortType) {
-            AlbumSortType.CREATE_DATE -> albumsByCreateDateAsc()
-            AlbumSortType.NAME -> albumsByNameAsc()
-            AlbumSortType.ARTIST -> albumByArtistAsc()
-            AlbumSortType.YEAR -> albumsByYearAsc()
-            AlbumSortType.SONG_COUNT -> albumsBySongCountAsc()
-            AlbumSortType.LENGTH -> albumsByLengthAsc()
-            AlbumSortType.PLAY_TIME -> albumsByPlayTimeAsc()
-        }.map { it.reversed(descending) }
-
-    fun albumsLiked(sortType: AlbumSortType, descending: Boolean) =
-        when (sortType) {
-            AlbumSortType.CREATE_DATE -> albumsLikedByCreateDateAsc()
-            AlbumSortType.NAME -> albumsLikedByNameAsc()
-            AlbumSortType.ARTIST -> albumLikeByArtistAsc()
-            AlbumSortType.YEAR -> albumsLikedByYearAsc()
-            AlbumSortType.SONG_COUNT -> albumsLikedBySongCountAsc()
-            AlbumSortType.LENGTH -> albumsLikedByLengthAsc()
-            AlbumSortType.PLAY_TIME -> albumsLikedByPlayTimeAsc()
-        }.map { it.reversed(descending) }
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE id = :id")
-    fun album(id: String): Flow<Album?>
-
-    @Transaction
-    @Query("SELECT * FROM album WHERE id = :albumId")
-    fun albumWithSongs(albumId: String): Flow<AlbumWithSongs?>
-
-    @Transaction
-    @RawQuery(observedEntities = [PlaylistEntity::class])
-    fun _getPlaylists(query: SupportSQLiteQuery): Flow<List<Playlist>>
-
-    private fun queryPlaylists(orderBy: String): SimpleSQLiteQuery {
-        return SimpleSQLiteQuery("""
-            SELECT p.*, COUNT(psm.playlistId) AS songCount
-            FROM playlist p
-                LEFT JOIN playlist_song_map psm ON p.id = psm.playlistId
-            WHERE p.bookmarkedAt IS NOT NULL OR p.isLocal = 1
-            GROUP BY p.id
-            ORDER BY $orderBy
-        """)
-    }
-
-    fun playlistsByCreateDateAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylists("p.rowId"))
-    fun playlistsByNameAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylists("p.name COLLATE NOCASE ASC"))
-    fun playlistsBySongCountAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylists("songCount ASC"))
-    fun editablePlaylistsByCreateDateAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylists("p.rowId ASC"))
-
-    fun playlists(sortType: PlaylistSortType, descending: Boolean) =
-        when (sortType) {
-            PlaylistSortType.CREATE_DATE -> playlistsByCreateDateAsc()
-            PlaylistSortType.NAME -> playlistsByNameAsc()
-            PlaylistSortType.SONG_COUNT -> playlistsBySongCountAsc()
-        }.map { it.reversed(descending) }
-
-    private fun queryPlaylistsWithDownloads(orderBy: String): SimpleSQLiteQuery {
-        return SimpleSQLiteQuery("""
-            SELECT p.*, COUNT(psm.playlistId) AS songCount
-            FROM playlist p
-                LEFT JOIN playlist_song_map psm ON p.id = psm.playlistId
-                INNER JOIN song s ON psm.songId = s.id and s.dateDownload IS NOT NULL
-            WHERE p.bookmarkedAt IS NOT NULL OR p.isLocal = 1
-            GROUP BY p.id
-            ORDER BY $orderBy
-        """)
-    }
-
-    fun playlistsWithDownloadsByCreateDateAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylistsWithDownloads("p.rowId"))
-    fun playlistsWithDownloadsByNameAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylistsWithDownloads("p.name COLLATE NOCASE ASC"))
-    fun playlistsWithDownloadsBySongCountAsc(): Flow<List<Playlist>> = _getPlaylists(queryPlaylistsWithDownloads("songCount ASC"))
-
-    fun playlistsWithDownloads(sortType: PlaylistSortType, descending: Boolean) =
-        when (sortType) {
-            PlaylistSortType.CREATE_DATE -> playlistsWithDownloadsByCreateDateAsc()
-            PlaylistSortType.NAME -> playlistsWithDownloadsByNameAsc()
-            PlaylistSortType.SONG_COUNT -> playlistsWithDownloadsBySongCountAsc()
-        }.map { it.reversed(descending) }
-
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(*) FROM playlist_song_map WHERE playlistId = playlist.id) AS songCount FROM playlist WHERE id = :playlistId")
-    fun playlist(playlistId: String): Flow<Playlist?>
-
-    @Query("SELECT songId from playlist_song_map WHERE playlistId = :playlistId AND songId IN (:songIds)")
-    fun playlistDuplicates(
-        playlistId: String,
-        songIds: List<String>,
-    ): List<String>
 
     @Transaction
     fun addSongToPlaylist(playlist: Playlist, songIds: List<String>) {
@@ -683,14 +99,7 @@ interface DatabaseDao {
         }
     }
 
-    @Transaction
-    @Query("SELECT *, (SELECT COUNT(*) FROM playlist_song_map WHERE playlistId = playlist.id) AS songCount FROM playlist WHERE browseId = :browseId")
-    fun playlistByBrowseId(browseId: String): Flow<Playlist?>
-
-    @Transaction
-    @Query("UPDATE playlist SET isLocal = 1 WHERE id = :playlistId")
-    fun playlistDesync(playlistId: String)
-
+    // region Search
     @Transaction
     @Query("SELECT * FROM song WHERE title LIKE '%' || :query || '%' AND inLibrary IS NOT NULL LIMIT :previewSize")
     fun searchSongs(query: String, previewSize: Int = Int.MAX_VALUE): Flow<List<Song>>
@@ -718,6 +127,7 @@ interface DatabaseDao {
     @Transaction
     @Query("SELECT *, (SELECT COUNT(*) FROM playlist_song_map WHERE playlistId = playlist.id) AS songCount FROM playlist WHERE name LIKE '%' || :query || '%' LIMIT :previewSize")
     fun searchPlaylists(query: String, previewSize: Int = Int.MAX_VALUE): Flow<List<Playlist>>
+    // endregion
 
     @Transaction
     @Query("SELECT * FROM event ORDER BY rowId DESC")
@@ -732,93 +142,8 @@ interface DatabaseDao {
     @Query("DELETE FROM search_history")
     fun clearSearchHistory()
 
-    @Query("UPDATE song SET totalPlayTime = totalPlayTime + :playTime WHERE id = :songId")
-    fun incrementTotalPlayTime(songId: String, playTime: Long)
-
-    @Query("UPDATE playCount SET count = count + 1 WHERE song = :songId AND year = :year AND month = :month")
-    fun incrementPlayCount(songId: String, year: Int, month: Int)
-
-    /**
-     * Increment by one the play count with today's year and month.
-     */
-    fun incrementPlayCount(songId: String) {
-        val time = LocalDateTime.now().atOffset(ZoneOffset.UTC)
-        var oldCount: Int
-        runBlocking {
-            oldCount = getPlayCountByMonth(songId, time.year, time.monthValue).first()
-        }
-
-        // add new
-        if (oldCount <= 0) {
-            insert(PlayCountEntity(songId, time.year, time.monthValue, 0))
-        }
-        incrementPlayCount(songId, time.year, time.monthValue)
-    }
-
-    @Transaction
-    fun toggleInLibrary(songId: String, inLibrary: LocalDateTime?) {
-        inLibrary(songId, inLibrary)
-        if (inLibrary == null) {
-            removeLike(songId)
-        }
-    }
-
-    @Query("UPDATE song SET inLibrary = :inLibrary WHERE id = :songId")
-    fun inLibrary(songId: String, inLibrary: LocalDateTime?)
-
-    @Query("UPDATE song SET liked = 0, likedDate = null WHERE id = :songId")
-    fun removeLike(songId: String)
-
-    @Query("UPDATE song SET inLibrary = null WHERE localPath = null")
-    fun disableInvalidLocalSongs()
-    @Query("UPDATE song SET inLibrary = null, localPath = null WHERE id = :songId")
-    fun disableLocalSong(songId: String)
-
-    @Query("""
-        SELECT * FROM song
-        WHERE localPath IN (
-            SELECT localPath
-            FROM song
-            GROUP BY localPath
-            HAVING COUNT(*) > 1
-        )
-        ORDER BY localPath
-    """)
-    fun duplicatedLocalSongs(): Flow<List<SongEntity>>
-
-    fun updateLocalSongPath(songId: String, inLibrary: LocalDateTime?, localPath: String?) {
-        if (localPath != null) {
-            updateLSP(songId, inLibrary, localPath)
-        }
-    }
-
-    /**
-     * DON'T USE THIS DIRECTLY, USE updateLocalSongPath(...) instead!
-     */
-    @Query("UPDATE song SET inLibrary = :inLibrary, localPath = :localPath WHERE id = :songId")
-    fun updateLSP(songId: String, inLibrary: LocalDateTime?, localPath: String)
-
     @Query("SELECT COUNT(1) FROM related_song_map WHERE songId = :songId LIMIT 1")
     fun hasRelatedSongs(songId: String): Boolean
-
-    @Query(
-        """
-        UPDATE playlist_song_map SET position = 
-            CASE 
-                WHEN position < :fromPosition THEN position + 1
-                WHEN position > :fromPosition THEN position - 1
-                ELSE :toPosition
-            END 
-        WHERE playlistId = :playlistId AND position BETWEEN MIN(:fromPosition, :toPosition) AND MAX(:fromPosition, :toPosition)
-    """
-    )
-    fun move(playlistId: String, fromPosition: Int, toPosition: Int)
-
-    @Query("DELETE FROM playlist_song_map WHERE playlistId = :playlistId")
-    fun clearPlaylist(playlistId: String)
-
-    @Query("SELECT * FROM artist WHERE name = :name")
-    fun artistByName(name: String): ArtistEntity?
 
     @Query("SELECT * FROM genre WHERE title = :name")
     fun genreByName(name: String): GenreEntity?
@@ -826,18 +151,6 @@ interface DatabaseDao {
     @Transaction
     @Query("SELECT * FROM genre WHERE title LIKE '%' || :query || '%' LIMIT :previewSize")
     fun genreByAproxName(query: String, previewSize: Int = Int.MAX_VALUE): Flow<List<GenreEntity>>
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(song: SongEntity): Long
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(artist: ArtistEntity)
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(album: AlbumEntity): Long
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(playlist: PlaylistEntity)
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     fun insert(genre: GenreEntity)
@@ -865,15 +178,6 @@ interface DatabaseDao {
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     fun insert(map: RelatedSongMap)
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(playCountEntity: PlayCountEntity): Long
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insert(queue: QueueEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insert(queueSong: QueueSongMap)
 
     @Transaction
     fun insert(mediaMetadata: MediaMetadata, block: (SongEntity) -> SongEntity = { it }) {
@@ -985,32 +289,8 @@ interface DatabaseDao {
     }
 
     @Update
-    fun update(song: SongEntity)
-
-    @Update
-    fun update(artist: ArtistEntity)
-
-    @Update
-    fun update(album: AlbumEntity)
-
-    @Update
-    fun update(playlist: PlaylistEntity)
-
-    @Update
     fun update(map: PlaylistSongMap)
 
-    @Update
-    fun update(queue: QueueEntity)
-
-    fun update(artist: ArtistEntity, artistPage: ArtistPage) {
-        update(
-            artist.copy(
-                name = artistPage.artist.title,
-                thumbnailUrl = artistPage.artist.thumbnail.resize(544, 544),
-                lastUpdateTime = LocalDateTime.now()
-            )
-        )
-    }
 
     @Transaction
     fun update(album: AlbumEntity, albumPage: AlbumPage) {
@@ -1053,20 +333,6 @@ interface DatabaseDao {
             ?.forEach(::insert)
     }
 
-    @Update
-    fun update(playlistEntity: PlaylistEntity, playlistItem: PlaylistItem) {
-        update(playlistEntity.copy(
-            name = playlistItem.title,
-            browseId = playlistItem.id,
-            isEditable = playlistItem.isEditable,
-            thumbnailUrl = playlistItem.thumbnail,
-            remoteSongCount = playlistItem.songCountText?.let { Regex("""\d+""").find(it)?.value?.toIntOrNull() },
-            playEndpointParams = playlistItem.playEndpoint?.params,
-            shuffleEndpointParams = playlistItem.shuffleEndpoint?.params,
-            radioEndpointParams = playlistItem.radioEndpoint?.params
-        ))
-    }
-
     @Transaction
     @Query("UPDATE song_artist_map SET artistId = :newId WHERE artistId = :oldId")
     fun updateSongArtistMap(oldId: String, newId: String)
@@ -1074,9 +340,6 @@ interface DatabaseDao {
     @Transaction
     @Query("UPDATE album_artist_map SET artistId = :newId WHERE artistId = :oldId")
     fun updateAlbumArtistMap(oldId: String, newId: String)
-
-    @Query("UPDATE song SET dateDownload = :dateDownload WHERE id = :songId")
-    suspend fun updateDownloadStatus(songId: String, dateDownload: LocalDateTime?)
 
     @Upsert
     fun upsert(map: SongAlbumMap)
@@ -1086,18 +349,6 @@ interface DatabaseDao {
 
     @Upsert
     fun upsert(format: FormatEntity)
-
-    @Delete
-    fun delete(song: SongEntity)
-
-    @Delete
-    fun delete(artist: ArtistEntity)
-
-    @Delete
-    fun delete(album: AlbumEntity)
-
-    @Delete
-    fun delete(playlist: PlaylistEntity)
 
     @Delete
     fun delete(playlistSongMap: PlaylistSongMap)
@@ -1111,24 +362,9 @@ interface DatabaseDao {
     @Delete
     fun delete(event: Event)
 
-    @Delete
-    fun delete(mq: QueueEntity)
-
     @Transaction
     @Query("DELETE FROM song_artist_map WHERE songId = :songID")
     fun unlinkSongArtists(songID: String)
-
-    @Transaction
-    @Query("DELETE FROM song WHERE isLocal = 1")
-    fun nukeLocalSongs()
-
-    @Transaction
-    @Query("DELETE FROM artist WHERE isLocal = 1")
-    fun nukeLocalArtists()
-
-    @Transaction
-    @Query("DELETE FROM album WHERE isLocal = 1")
-    fun nukeLocalAlbums()
 
     @Transaction
     @Query("DELETE FROM genre WHERE isLocal = 1")
@@ -1158,7 +394,6 @@ interface DatabaseDao {
     /**
      * Queueboard
      */
-
     @Transaction
     fun saveQueue(mq: MultiQueueObject) {
         if (mq.queue.isEmpty() || mq.unShuffled.isEmpty()) {
@@ -1198,75 +433,6 @@ interface DatabaseDao {
             )
         }
     }
-
-    @Transaction
-    fun updateQueue(mq: MultiQueueObject) {
-        update(
-            QueueEntity(
-                id = mq.id,
-                title = mq.title,
-                shuffled = mq.shuffled,
-                queuePos = mq.queuePos,
-                index = mq.index,
-                playlistId = mq.playlistId
-            )
-        )
-    }
-
-    @Transaction
-    fun updateAllQueues(mqs: List<MultiQueueObject>) {
-        CoroutineScope(Dispatchers.IO).launch {
-            QueueBoard.mutex.withLock { // possible ConcurrentModificationException
-                mqs.forEach { updateQueue(it) }
-            }
-        }
-    }
-
-    @Query("SELECT * from queue ORDER BY `index`")
-    fun getAllQueues(): Flow<List<QueueEntity>>
-
-    @Transaction
-    @RewriteQueriesToDropUnusedColumns
-    @Query("SELECT * from queue_song_map JOIN song ON queue_song_map.songId = song.id WHERE queueId = :queueId AND shuffled = 1")
-    fun getQueueSongs(queueId: Long): Flow<List<Song>>
-
-    @Transaction
-    @RewriteQueriesToDropUnusedColumns
-    @Query("SELECT * from queue_song_map JOIN song ON queue_song_map.songId = song.id WHERE queueId = :queueId AND shuffled = 0")
-    fun getQueueSongsUnshuffled(queueId: Long): Flow<List<Song>>
-
-    fun readQueue(): List<MultiQueueObject> {
-        val resultQueues = ArrayList<MultiQueueObject>()
-        val queues = runBlocking { getAllQueues().first() }
-
-        queues.forEach { queue ->
-            val shuffledSongs = runBlocking { getQueueSongs(queue.id).first() }
-            val unshuffledSongs = runBlocking { getQueueSongsUnshuffled(queue.id).first() }
-
-            resultQueues.add(
-                MultiQueueObject(
-                    id = queue.id,
-                    title = queue.title,
-                    queue = shuffledSongs.map { it.toMediaMetadata() }.toMutableList(),
-                    unShuffled = unshuffledSongs.map { it.toMediaMetadata() }.toMutableList(),
-                    shuffled = queue.shuffled,
-                    queuePos = queue.queuePos,
-                    index = queue.index
-                )
-            )
-        }
-
-        return resultQueues
-    }
-
-    @Query("DELETE FROM queue")
-    fun deleteAllQueues()
-
-    @Query("DELETE FROM queue_song_map WHERE queueId = :id")
-    fun deleteAllQueueSongs(id: Long)
-
-    @Query("DELETE FROM queue WHERE id = :id")
-    fun deleteQueue(id: Long)
 
     /**
      * WARNING: This removes all queue data and re-adds the queue. Did you mean to use updateQueue()?
